Remove redundant linspace dtype
Agent-Logs-Url: https://github.com/SamuelDegelia-NOAA/RRFS_GETKF_PARALLEL/sessions/83d6ea6a-4119-4105-a199-dfe11bc5d6c9

+++ util/plot_bias_rmse_profile.py
@@ -205,8 +205,7 @@ def build_pressure_bins():
     first_center = PRESSURE_MIN_HPA + half_width
     num_bins = int((PRESSURE_MAX_BIN_CENTER_HPA - first_center)
                    / PRESSURE_BIN_WIDTH_HPA) + 1
-    centers = np.linspace(first_center, PRESSURE_MAX_BIN_CENTER_HPA,
-                          num=num_bins, dtype=float)
+    centers = np.linspace(first_center, PRESSURE_MAX_BIN_CENTER_HPA, num=num_bins)
     bins = []
     for i, center in enumerate(centers):
         low = center - half_width
